% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{cleanOutfile}
\alias{cleanOutfile}
\title{Create outdir if not exist, and remove the old file if needed.}
\usage{
cleanOutfile(outf, remove = TRUE)
}
\arguments{
\item{outf}{string}

\item{remove}{bool, default is TRUE}
}
\value{
None
}
\description{
Create outdir if not exist, and remove the old file if needed.
}
