% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{colVector}
\alias{colVector}
\title{Create a color panel
Ref: From SnapATAC}
\usage{
colVector(num.color = 60, type = c("qual", "div", "seq"))
}
\arguments{
\item{num.color}{Number of diffenrent colors to return}

\item{type}{Type of colors to generate}
}
\description{
Create a color panel
Ref: From SnapATAC
}
