% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{createColorPanel}
\alias{createColorPanel}
\title{Create a color panel
Ref: From SnapATAC}
\usage{
createColorPanel(num.color)
}
\arguments{
\item{num.color}{Number of diffenrent colors to return}
}
\description{
Create a color panel
Ref: From SnapATAC
}
