% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{fastGetTSSEnrichmentMultiThreads}
\alias{fastGetTSSEnrichmentMultiThreads}
\title{Get TSS Enrichment scores}
\usage{
fastGetTSSEnrichmentMultiThreads(
  TSS,
  barcodes,
  rawH5File,
  nChunkInRawH5File = 3,
  window = 101,
  norm = 100,
  flank = 2000,
  minNorm = 0.2,
  maxFragSize = NULL,
  sampleName = NULL,
  nthread = 2
)
}
\value{
List two element: TSSE, TSSReads
}
\description{
Ref: ArchR
}
