% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{fastGetTSSEnrichmentSingleThread}
\alias{fastGetTSSEnrichmentSingleThread}
\title{Get TSS Enrichment scores}
\usage{
fastGetTSSEnrichmentSingleThread(
  TSS,
  barcodes,
  rawH5File,
  window = 101,
  norm = 100,
  flank = 2000,
  minNorm = 0.2,
  maxFragSize = NULL,
  sampleName = NULL
)
}
\value{
List two element: TSSE, TSSReads

dataframe with cols, barcode,TSSE,TSSRead
}
\description{
Ref: ArchR
}
