% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{fastH5Read}
\alias{fastH5Read}
\title{Fast read H5 file
Ref: ArchR .h5read}
\usage{
fastH5Read(
  file = NULL,
  name = NULL,
  method = "fast",
  index = NULL,
  start = NULL,
  block = NULL,
  count = NULL
)
}
\value{
results return by h5read
}
\description{
Fast read H5 file
Ref: ArchR .h5read
}
