% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGenome.R
\name{filterChrGR}
\alias{filterChrGR}
\title{filter ChrGR}
\usage{
filterChrGR(gr = NULL, remove = NULL, understcore = TRUE)
}
\arguments{
\item{gr}{GRanges}

\item{remove}{charactervector extra seqlevels to be removed}

\item{underscore}{bool remove seqlevels containing an underscore}

\item{standard}{bool keep all the standard genomes}
}
\value{
gr GRanges
}
\description{
Ref: ArchR filterChrGR
}
