% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGenome.R
\name{getAnnotFromArchRData}
\alias{getAnnotFromArchRData}
\title{get gene or genome Annotation from ArchR data.}
\usage{
getAnnotFromArchRData(tag, genome)
}
\arguments{
\item{tag}{string, gene or genome}

\item{genome}{string genomenm}
}
\value{
annotation SimpleGenomicRangesList
}
\description{
Ref: ArchR getArchRGenome
}
\examples{
getAnnotFromArchRData(tag = "gene", genome = "mm10")
}
