% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGenome.R
\name{getBlacklist}
\alias{getBlacklist}
\title{get black list}
\usage{
getBlacklist(genome)
}
\arguments{
\item{genome}{str}
}
\value{
GRanges
}
\description{
Return GRanges. Empty GRanges will be return if no name matches the genome or network issusue.
Ref: ArchR .getBlacklist
}
