% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getFragsOfAChrFromRawH5File}
\alias{getFragsOfAChrFromRawH5File}
\title{Get fragments from a given chromsome in rawH5File.
Ref: ArchR .getFragsFromArrow}
\usage{
getFragsOfAChrFromRawH5File(
  rawH5File,
  chr = "chr1",
  sampleName = NULL,
  barcodes = NULL,
  verbose = TRUE
)
}
\arguments{
\item{barcodes}{vector of strings, make sure it's not a data.frame with one column.}

\item{verbose}{bool, print logs when set is TRUE, default is TRUE.}
}
\value{
fragments IRanges
}
\description{
Get fragments from a given chromsome in rawH5File.
Ref: ArchR .getFragsFromArrow
}
