% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneMatrix.R
\name{getGmatFromSnap}
\alias{getGmatFromSnap}
\title{Get Gmat from Snap object}
\usage{
getGmatFromSnap(
  snapFile,
  geneBedFile = NULL,
  ibarcode = NULL,
  blacklistBedFile = NULL,
  outfile = NULL,
  compressLevel = 9
)
}
\arguments{
\item{snapFile}{string}

\item{geneBedFile}{string, default is NULL. Only used for mapping with blacklist}

\item{ibarcode}{vector of string, the barcodes we want to, we assume all of them are in
this snapFile. So use this carefully.}

\item{blacklistBedFile}{string, bed file of the blacklist, smmtools save one under data dir for mm10.}

\item{outfile}{string, default is NULL, otherwise please end of .h5, and the function will save gmat.
if the file already exists, the function will delete it.}

\item{compressLevel}{integer, default is 9, used for rhdf5.}
}
\value{
sparseMatrix, cell by feature, with both row and colname and ordered by ibarcode if provided
}
\description{
Get Gmat from Snap object
}
