% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getiFromSparseMatrix}
\alias{getiFromSparseMatrix}
\title{Get i from sparseMatrix
i index starts froms 1.
Ref: https://stackoverflow.com/questions/21099612/extract-i-and-j-from-a-sparse-matrix}
\usage{
getiFromSparseMatrix(sm)
}
\arguments{
\item{sm}{sparseMatrix}
}
\value{
vector of integer
}
\description{
Get i from sparseMatrix
i index starts froms 1.
Ref: https://stackoverflow.com/questions/21099612/extract-i-and-j-from-a-sparse-matrix
}
