% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getIndex}
\alias{getIndex}
\title{Get index in the subject vector for the query vector}
\usage{
getIndex(query, subject)
}
\arguments{
\item{query}{vector, same type as in subject}

\item{subject}{vector, same type as in query}
}
\value{
integer vector, index in subject for the query vector
}
\description{
Get index in the subject vector for the query vector
}
