% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getjFromSparseMatrix}
\alias{getjFromSparseMatrix}
\title{Get j from sparseMatrix
j index starts froms 1.
Ref: https://stackoverflow.com/questions/21099612/extract-i-and-j-from-a-sparse-matrix}
\usage{
getjFromSparseMatrix(sm)
}
\arguments{
\item{sm}{sparseMatrix}
}
\value{
vector of integer
}
\description{
Get j from sparseMatrix
j index starts froms 1.
Ref: https://stackoverflow.com/questions/21099612/extract-i-and-j-from-a-sparse-matrix
}
