% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{getNfragmentPerBarcode}
\alias{getNfragmentPerBarcode}
\title{get numebr of fragments per barcode
Ref: ArchR}
\usage{
getNfragmentPerBarcode(chrRegions, rawH5File)
}
\arguments{
\item{chrRegions}{array of characters, chrRegion names in rawH5File}

\item{rawH5File}{string}
}
\value{
tibble shape = \link{length(barcodes), 2}, colnames = c("barcode", "count")
}
\description{
get numebr of fragments per barcode
Ref: ArchR
}
