% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGenome.R
\name{getTSSFromTxDb}
\alias{getTSSFromTxDb}
\title{get TSS from TxDb}
\usage{
getTSSFromTxDb(TxDb)
}
\arguments{
\item{TxDb}{TxDbobjet}
}
\value{
TSS GenomicRanges
}
\description{
Ref: ArchR createGeneAnnotation
}
\examples{
library(TxDb.Mmusculus.UCSC.mm10.knownGene)
TxDb <- TxDb.Mmusculus.UCSC.mm10.knownGene
TSS <- getTSSFromTxDb(TxDb)
}
