% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{getUmapLikeArchR}
\alias{getUmapLikeArchR}
\title{Umap based on ArchR default parameters}
\usage{
getUmapLikeArchR(
  mapSnapATAC,
  nPC = 30,
  seed = 10,
  weightDimReduct = FALSE,
  umapNeighbors = 40,
  umapMindist = 0.4,
  umapMetric = "cosine",
  umapVerbose = FALSE,
  umapFastSGD = TRUE,
  umapNthread = 2
)
}
\description{
Umap based on ArchR default parameters
}
