% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getvalFromSparseMatrix}
\alias{getvalFromSparseMatrix}
\title{Get x/values from sparseMatrix}
\usage{
getvalFromSparseMatrix(sm)
}
\arguments{
\item{sm}{sparseMatrix}
}
\value{
numeric vector
}
\description{
Ref: https://stackoverflow.com/questions/21099612/extract-i-and-j-from-a-sparse-matrix
}
