% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGenome.R
\name{installGenomeRelatedDatabase}
\alias{installGenomeRelatedDatabase}
\title{Install genome/TxDb/OrgDb}
\usage{
installGenomeRelatedDatabase(genome, dbnm)
}
\arguments{
\item{genome}{str}

\item{dbnm}{str}
}
\value{
name str
}
\description{
Install the genome from BiocManager if the corresponding BSgenome does not exists.
Ref: ArchR addArchRGenome
}
