% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_utils.R
\name{load_sparse_matrix}
\alias{load_sparse_matrix}
\title{Load sparse matrix from mat file}
\usage{
load_sparse_matrix(matfile, ...)
}
\arguments{
\item{matfile}{str}
}
\value{
sparseMatrix
}
\description{
Each row of matfile: barcode\link{,genome_bin_index:counts}
}
