% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileMatrix.R
\name{loadBmatFromFile}
\alias{loadBmatFromFile}
\title{Load SnapATAC Bmat from the h5 file generated by getBmatFromSnap}
\usage{
loadBmatFromFile(bmatFile)
}
\arguments{
\item{bmatFile}{string, should be h5 file.}
}
\value{
sparseMatrix, cell by bins with names
}
\description{
Load SnapATAC Bmat from the h5 file generated by getBmatFromSnap
}
