% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneMatrix.R
\name{loadGeneMatrix}
\alias{loadGeneMatrix}
\title{Load sparse matrix of gene matrix with both row and col names}
\usage{
loadGeneMatrix(geneMatrixH5File)
}
\arguments{
\item{geneMatrixH5File}{string, h5 file to read}
}
\description{
Load sparse matrix of gene matrix with both row and col names
}
