% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneMatrix.R
\name{loadSnapATACGmatFromFile}
\alias{loadSnapATACGmatFromFile}
\title{Load SnapATAC Gmat from the h5 file.}
\usage{
loadSnapATACGmatFromFile(gmatFile)
}
\arguments{
\item{gmatFile}{string}
}
\value{
sparseMatrix, cell by gene with names
}
\description{
Load SnapATAC Gmat from the h5 file.
}
