% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{loadSparseMatrixFromMatTextFile}
\alias{loadSparseMatrixFromMatTextFile}
\title{Load sparse matrix from mat file}
\usage{
loadSparseMatrixFromMatTextFile(matfile, ...)
}
\arguments{
\item{matfile}{str}

\item{...}{for Matrix::sparseMatrix usage}
}
\value{
sparseMatrix
}
\description{
Each row of matfile: barcode\link{,genome_bin_index:counts}
}
