% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plot2DEmbed}
\alias{plot2DEmbed}
\title{Ref: From SnapATAC}
\usage{
plot2DEmbed(
  dmat,
  point.size = 1,
  point.shape = 19,
  point.alpha = 0.8,
  clusterLabel = NULL,
  text.add = TRUE,
  text.size = 1,
  text.color = "black",
  text.halo.add = TRUE,
  text.halo.color = "white",
  text.halo.width = 0.2,
  legend.add = FALSE,
  legend.pos = c("bottomleft", "bottom", "left", "topleft", "top", "topright", "right",
    "center"),
  legend.text.size = 1,
  legend.text.color = "black",
  down.sample = 10000,
  pdf.file.name = NULL,
  pdf.width = 7,
  pdf.height = 7,
  myxlims = NULL,
  myylims = NULL,
  ...
)
}
\description{
Ref: From SnapATAC
}
