% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{plotDimReductElbow}
\alias{plotDimReductElbow}
\title{Plot correlations between adjacent two principle components.
Refer: SnapATAC plotDimReductElbow}
\usage{
plotDimReductElbow(
  sdev,
  outpdf = NULL,
  point_size = 1.5,
  point_shape = 19,
  point_color = "red",
  point_alpha = 1,
  height = 7,
  width = 7,
  ...
)
}
\arguments{
\item{sdev}{numeric vector, standard deviation for PCs.}

\item{outpdf}{string, output of pdf file, default is NULL.}

\item{point_size}{double, default 1.5}

\item{point_shape}{integer, default 19}

\item{point_color}{string, default "red"}

\item{point_alpha}{double, default 1}

\item{height}{double, height of pdf, default 7}

\item{width, }{double, width of pdf, default 7}

\item{...}{parameters, used for plot.}
}
\value{
None
}
\description{
Plot correlations between adjacent two principle components.
Refer: SnapATAC plotDimReductElbow
}
