% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{plotDimReductPairwise}
\alias{plotDimReductPairwise}
\title{Plot correlations between adjacent two principle components.
Refer: SnapATAC plotDimReductPW}
\usage{
plotDimReductPairwise(
  smat,
  outpdf = NULL,
  ndim = 50,
  point_size = 0.5,
  point_color = "grey",
  point_shape = 19,
  point_alpha = 0.5,
  nsample = 10000,
  height = 7,
  width = 7,
  ...
)
}
\arguments{
\item{smat}{cell by feature}

\item{outpdf}{string, output of pdf file, default is NULL.}

\item{ndim}{integer, max dims to plot, default is 50}

\item{point_size}{double, default 0.5}

\item{point_color}{string, default "grey"}

\item{point_shape}{integer, default 19}

\item{point_alpha}{double, default 0.5}

\item{nsample}{integer, max number of random cells to use, default 10000}

\item{height}{double, height of pdf, default 7}

\item{width, }{double, width of pdf, default 7}

\item{...}{parameters, used for plot.}
}
\value{
value returned by graphics::par()
}
\description{
Plot correlations between adjacent two principle components.
Refer: SnapATAC plotDimReductPW
}
