% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{read_columns}
\alias{read_columns}
\title{Read file by columns}
\usage{
read_columns(filenm, cols = c(1), sep = ",", ...)
}
\arguments{
\item{filenm}{str}

\item{cols}{vector\link{Integer}}

\item{sep}{str}

\item{...}{others for read.csv}
}
\value{
data.frame from read.csv
}
\description{
Ref: https://www.gastonsanchez.com/visually-enforced/how-to/2012/06/23/Read-file-by-columns/
}
