% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{removeSampleName}
\alias{removeSampleName}
\title{Remove sampleName in barcodes if they have}
\usage{
removeSampleName(barcodes, sep = "#")
}
\arguments{
\item{barcodes}{vector of strings}

\item{sep}{string, symbol used to separate sampleName and barcode, default is '#'}
}
\description{
Remove sampleName in barcodes if they have
}
