% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{runJaccard2}
\alias{runJaccard2}
\title{Get Jaccard Matrix
Ref: SnapATAC runJaccard2}
\usage{
runJaccard2(xi, xj)
}
\arguments{
\item{xi}{matrix/sparseMatrix, cell by feature, multi-hot representation}

\item{xj}{matrix/sparseMatrix, cell by feature, multi-hot representation}
}
\value{
dense matrix, Jaccard Matrix
}
\description{
Get Jaccard Matrix
Ref: SnapATAC runJaccard2
}
