% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{runKNN}
\alias{runKNN}
\title{Simplified runKNN in SnapATAC}
\usage{
runKNN(smat, k = 20, nn_eps = 0)
}
\arguments{
\item{nn_eps}{Error bound when performing nearest neighbor seach using RANN.
default of 0.0 implies exact nearest neighbor search}
}
\description{
Simplified runKNN in SnapATAC
}
