% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{runLeiden}
\alias{runLeiden}
\title{Run Leiden algorithm as graph-based clustering.}
\usage{
runLeiden(
  kmat,
  path_to_python = NULL,
  reso = 0.8,
  seed = 10,
  partitionType = "RB"
)
}
\arguments{
\item{kmat}{sparseMatrix, KNN matrix generated by runKNN.}

\item{reso}{double, resolution param in Leiden default is 0.8}

\item{seed}{integer, used for Leiden}

\item{partitionType}{string, used for Leiden, default is "RB"}

\item{path_to_ptyhon}{string}
}
\value{
vector of factor, cluster index for cells
}
\description{
Run Leiden algorithm as graph-based clustering.
}
