% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublet.R
\name{runScrublet}
\alias{runScrublet}
\title{Doublet removement using Scrublet}
\usage{
runScrublet(
  bmat,
  path_to_python,
  expected_doublet_rate = 0.6,
  min_counts = 3,
  min_cells = 5L,
  min_conv_pctl = 85,
  n_pc = 30L
)
}
\value{
list, five fields
\enumerate{
\item threshold based on GMM for simulation scores \link{5}
\item threshold estimated from Scrublets
\item probs of real data based on GMM
\item doublet scores of real data based on Scrublets
\item doublet scores of simulation data based on Scrublets
}
}
\description{
Need smmutil
}
