% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{runUMAP}
\alias{runUMAP}
\title{Umap based on ArchR and Seuratdefault parameters}
\usage{
runUMAP(
  dmat,
  sdev = NULL,
  nPC = 30,
  outpdf = NULL,
  width = 7,
  height = 7,
  xlab = "UMAP-1",
  ylab = "UMAP-1",
  title = NULL,
  color = NULL,
  ncomp = 2,
  seed = 10,
  weightDimReduct = FALSE,
  umapNeighbors = 30,
  umapMindist = 0.3,
  umapMetric = "cosine",
  umapVerbose = FALSE,
  umapFastSGD = TRUE,
  umapNthread = 1,
  a = NULL,
  b = NULL,
  init = "spectral",
  retnn = FALSE,
  retmodel = FALSE
)
}
\arguments{
\item{dmat}{matrix, cell by feature}

\item{sdev}{vector, scaling factors for dims, default is NULL}

\item{nPC}{integer, dims till the nPC-th dim will be used for UMAP, default is 30.
When nfeature in mat is less than nPC, nfeature will be used.}

\item{outpdf}{string, default is NULL}

\item{width}{double, for outpdf, default is 7}

\item{height}{double, for outpdf, default is 7}

\item{xlab}{string, default is "UMAP-1"}

\item{ylab}{string, default is "UMAP-2"}

\item{title}{string, default is NULL}

\item{color}{vector of string or factor, default is NULL}

\item{ncomp}{integer, defaul is 2}

\item{seed}{integer, default is 10}

\item{weightDimReduct}{bool, default is FALSE, together used with sdev.}

\item{umapNeighbors}{integer, default is 30
In umot::umap, default is 15.}

\item{umapMindist}{double, default is 0.3
In umot::umap, default is 0.01}

\item{umapMetric}{string, default is "cosine"
NOTE: "euclidean" is the default in uwot::umap}

\item{umapVerbose}{bool, default is FALSE}

\item{umapFastSGD}{bool, default is TRUE}

\item{umapNthread}{integer, default is 1}

\item{a}{double, default is NULL}

\item{b}{double, default is NULL
NOTE: a = 1.8956, b = 0.8006 is the default UMAP}

\item{init}{string, default is "special"
NOTE: Use init = "spca" rather than init = "spectral" for better reproducibility
(although the latter is the default and preferred method for UMAP initialization).}

\item{retnn}{bool, default is FALSE}

\item{retmodel}{bool, default is FALSE
Set both retnn and retmodel as TRUE to save model}
}
\value{
matrix, cell by ncomp
}
\description{
Use uwot pacakge since it's more stable in R.
Uwot::umap default parameters are from ArchR.
}
\details{
Ref: https://github.com/jlmelville/uwot
Ref: https://satijalab.org/seurat/reference/runumap
}
