% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{sampleBasedOnDepth2}
\alias{sampleBasedOnDepth2}
\title{Sampling cells based on their sequencing depths without replace.}
\usage{
sampleBasedOnDepth2(log10UMI, n, seed = 2022)
}
\arguments{
\item{n}{integer}

\item{bmat}{Matrix, cell by feature}
}
\value{
vector of integers, index for the sampled cells
with length of min(n, ncells)
}
\description{
Sampling cells based on their sequencing depths without replace.
}
