% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchCorrection.R
\name{SnapATAC_runHarmony}
\alias{SnapATAC_runHarmony}
\title{Perform batch effect using harmony}
\usage{
SnapATAC_runHarmony(
  mapSnapATAC,
  nPC,
  meta_data,
  vars_use = NULL,
  weightDimReduct = FALSE,
  ...
)
}
\arguments{
\item{meta_data}{Either (1) Dataframe with variables to integrate or (2) vector with labels.}

\item{vars_use}{If meta_data is dataframe, this defined which variable(s) to remove (character vector).}

\item{...}{Paramters passed to Harmony}
}
\value{
Matrix, cell by updated principle components
}
\description{
Perform batch effect using harmony
}
