% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPalettes.R
\docType{data}
\name{SnapATACPalette}
\alias{SnapATACPalette}
\title{The color palettes from SnapATAC}
\format{
An object of class \code{character} of length 89.
}
\usage{
SnapATACPalette
}
\description{
The color palettes from SnapATAC
}
\keyword{datasets}
