% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrateATACWithRNA.R
\name{snapGmat2Seurat}
\alias{snapGmat2Seurat}
\title{Conver snap object to Seurat by gmat}
\usage{
snapGmat2Seurat(
  snap,
  eigDims = 1:50,
  assay = "GeneScore",
  pcaPrefix = "SnapATAC_"
)
}
\arguments{
\item{snap}{snap object defined in SnapATAC package
snap@gmat must be not empty; snap@smat@dmat should not be empty}

\item{eigDims}{vector, used for choosing PCA components, default 1:50}

\item{assay}{characters, name used in Seurat object}

\item{pcaPrefix}{characters, default "SnapATAC_"}
}
\value{
Seurat object
}
\description{
Conver snap object to Seurat by gmat
}
