% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGenome.R
\name{subsetGeneAnnoByGenomeAnno}
\alias{subsetGeneAnnoByGenomeAnno}
\title{subset GeneAnnotation by GenomeAnnotation}
\usage{
subsetGeneAnnoByGenomeAnno(geneAnnotation, genomeAnnotation)
}
\arguments{
\item{geneAnnotation}{SimpleGenomicRangesList for genes}

\item{genomeAnnotation}{SimpleGenomicRangesList for genomic}
}
\value{
geneAnnotation SimpleGenomicRangesList limited by genomicAnnotation
}
\description{
Ref: ArchR .validGeneAnnoByGenomeAnno
}
