% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{subsetSeqnamesGR}
\alias{subsetSeqnamesGR}
\title{get subset of GR}
\usage{
subsetSeqnamesGR(gr, names)
}
\arguments{
\item{gr}{GenomeRanges}

\item{names}{chromosome}
}
\value{
gr2 subset of gr whose seqnames limited in names
}
\description{
Ref: ArchR .subsetSeqnamesGR
}
