% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{sumFragmentSingleThread}
\alias{sumFragmentSingleThread}
\title{Transform original tabix File to H5File, then get barcodes-related fragments stats.}
\usage{
sumFragmentSingleThread(
  tabixFile,
  rawH5File,
  sumFragFile,
  coverH5File,
  genome,
  sampleName = NULL,
  barcodes = NULL,
  nChunk = 3,
  tsseWindow = 101,
  tsseNorm = 100,
  tsseFlank = 2000,
  tsseMinNorm = 0.2,
  tsseMaxFragSize = NULL
)
}
\arguments{
\item{tabixFile}{string, 10x cell ranger result tab/tab.gz file name}

\item{coverH5File}{bool, whether to cover the rawH5File if exists}

\item{genome}{string, namely mm9, mm10, hg19 or hg38}

\item{sampleName}{string, name for the tabix file}

\item{barcodes}{vector of string, limite fragment on specific barcodes}

\item{nChunk}{integer, partition each chrom into nChunk pieces}

\item{tsseWindow}{integer}

\item{tsseNorm}{interger}

\item{tsseFlank}{integer}

\item{tsseMinNorm}{float}

\item{tsseMaxFragSize}{integer}

\item{outfilenm}{string, output file name, if NULL, a default one will be genereated.}
}
\value{
data.frame with cols: barcode, nUniqFrag, TSSE, TSSReads
}
\description{
\itemize{
\item Ref: ArchR
\item Method is run on single core, but the speed should be OK.
\item No filtering barcodes here unless we provide barcode array ourselves.
\item Side Effect:
\itemize{
\item output rawH5File
\item output sumFragment File (descreasingly ordered by nUniqFrag)
}
}
}
