% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{t.sparseMatrix}
\alias{t.sparseMatrix}
\title{Transpose a sparseMatrix
A method for the generics function "t"
DEPRECATED: we can directly use Matrix::t.}
\usage{
t.sparseMatrix(sm)
}
\arguments{
\item{sm}{sparseMatrix}
}
\value{
sparseMatrix
}
\description{
Transpose a sparseMatrix
A method for the generics function "t"
DEPRECATED: we can directly use Matrix::t.
}
