% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabix.R
\name{tabixToH5SingleThread}
\alias{tabixToH5SingleThread}
\title{Read fragment information}
\usage{
tabixToH5SingleThread(
  tabixFile,
  tileChromSizes,
  sampleName = NULL,
  barcodes = NULL,
  outH5File = tempfile(pattern = paste0("tmp-", sampleName, ".h5"), tmpdir = tempdir())
)
}
\arguments{
\item{tabixFile}{string fragment file name, only tab/tab.gz file (.tabix should be in the same folder)}

\item{tileChromSizes}{GenomeRanges, tiled ChromSizes}

\item{sampleName}{string the sample corresponding to the fragment file}

\item{barcodes}{vector of string, optional, barcodes we want}

\item{outH5File}{string filename of the output h5file}

\item{nChunk}{integer partition each chrom into nChunk pieces}
}
\value{
outputH5File string name of the output file
}
\description{
Ref: ArchR .tabixTotmp
}
