% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{tileChrom}
\alias{tileChrom}
\title{tile ChromeSize
Ref: ArchR}
\usage{
tileChrom(chromSizes, nChunk = 3)
}
\arguments{
\item{chromSizes}{GenomeRanges}

\item{nChunk}{integer, default is 3}
}
\value{
tileChromSizes GenomeRanges
}
\description{
tile ChromeSize
Ref: ArchR
}
