% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getFragsOfAChrFromRawH5File}
\alias{getFragsOfAChrFromRawH5File}
\title{Get fragments from a given chromsome in rawH5File.
Ref: ArchR .getFragsFromArrow}
\usage{
getFragsOfAChrFromRawH5File(
  rawH5File,
  chr = "chr1",
  sampleName = NULL,
  nChunk = 3
)
}
\value{
fragments IRanges
}
\description{
Get fragments from a given chromsome in rawH5File.
Ref: ArchR .getFragsFromArrow
}
