% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneMatrix.R
\name{getGeneMatrix}
\alias{getGeneMatrix}
\title{Generate Gene Matrix by directly mapping fragments onto genes.}
\usage{
getGeneMatrix(
  rawH5File,
  outdir,
  outfilenm,
  genes = NULL,
  genenms = NULL,
  barcodes = NULL,
  genome = "mm10",
  sampleName = NULL,
  excludeChr = c("chrM"),
  compressLevel = 9
)
}
\description{
Generate Gene Matrix by directly mapping fragments onto genes.
}
