% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getiFromSparseMatrix}
\alias{getiFromSparseMatrix}
\title{Ref: https://stackoverflow.com/questions/21099612/extract-i-and-j-from-a-sparse-matrix}
\usage{
getiFromSparseMatrix(sm)
}
\description{
Ref: https://stackoverflow.com/questions/21099612/extract-i-and-j-from-a-sparse-matrix
}
