% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublet.R
\name{getScrubletThresholdByMixEM}
\alias{getScrubletThresholdByMixEM}
\title{Another way to estimate the threshold for Scrublet
Method used by Yang Li in CEMBA 1.0.}
\usage{
getScrubletThresholdByMixEM(scrubletSimScores, defaultCutoff, prob = 0.5)
}
\arguments{
\item{scrubletSimScores}{numeric vector}

\item{defaultCutoff}{double, if mixEM failed, use this as threschold.}

\item{prob}{double, prob of a scrublet score treated as doublet, default is 0.5}
}
\value{
double
}
\description{
Another way to estimate the threshold for Scrublet
Method used by Yang Li in CEMBA 1.0.
}
