% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileMatrix.R
\name{getTileMatrix}
\alias{getTileMatrix}
\title{Generate matrix of couts for each tile per cell}
\usage{
getTileMatrix(
  rawH5File,
  outdir,
  outfilenm,
  genome = "mm10",
  sampleName = NULL,
  barcodes = NULL,
  tileSize = 5000,
  excludeChr = c("chrM", "chrY")
)
}
\arguments{
\item{outfilenm}{string, end with .h5}

\item{barcodes}{vector of string, no need to append sampleName, will do that in this code block}
}
\description{
Ref: ArchR
}
