% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileMatrix.R
\name{loadTileMatrix}
\alias{loadTileMatrix}
\title{load TileMatrix into a sparseMatrix
Ref: ArchR}
\usage{
loadTileMatrix(tileMatrixFile, barcodes = NULL, binarize = FALSE)
}
\arguments{
\item{barcodes}{vector of string, choose the columns of tilematrix, useful after doublet removement}

\item{binarize}{bool, default is FALSE}
}
\description{
load TileMatrix into a sparseMatrix
Ref: ArchR
}
