% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileMatrix.R
\name{loadTileMatrix}
\alias{loadTileMatrix}
\title{load TileMatrix into a sparseMatrix
Ref: ArchR}
\usage{
loadTileMatrix(tileMatrixFile, barcodes = NULL)
}
\arguments{
\item{tileMatrixFile}{string}

\item{barcodes}{vector of string, if not null, filter tilematrix by barcodes, default is NULL.
We assume barcodes are all in the tileMatrixFile. NOTE: No check for this.}
}
\value{
sparseMatrix, feature by cell, both row and colnames, ordered by barcodes if provides.
}
\description{
load TileMatrix into a sparseMatrix
Ref: ArchR
}
