% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublet.R
\name{runScrublet}
\alias{runScrublet}
\title{Doublet removement using Scrublet}
\usage{
runScrublet(
  bmat,
  path_to_python,
  expected_doublet_rate = 0.6,
  min_counts = 3,
  min_cells = 5L,
  min_conv_pctl = 85,
  n_pc = 30L
)
}
\description{
Doublet removement using Scrublet
}
