% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{sampleBasedOnDepth}
\alias{sampleBasedOnDepth}
\title{Sampling cells based on their sequencing depths without replace.}
\usage{
sampleBasedOnDepth(bmat, n)
}
\arguments{
\item{bmat}{Matrix, cell by feature}

\item{n}{integer}
}
\value{
vector of integers, index for the sampled cells
with length of min(n, ncells)
}
\description{
Sampling cells based on their sequencing depths without replace.
}
