% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{simplifyGRSeqnames}
\alias{simplifyGRSeqnames}
\title{Remove chr in GRSeqnames}
\usage{
simplifyGRSeqnames(gr)
}
\arguments{
\item{gr}{GenomeRanges}
}
\value{
gr2 GenomeRanges simplified
}
\description{
Ref: ArchR .convertGRSeqnames
}
