% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureFilter.R
\name{SnapATAC_BinarizeBmat}
\alias{SnapATAC_BinarizeBmat}
\title{Binarize Bmat according SnapATAC}
\usage{
SnapATAC_BinarizeBmat(bmat, z_threshold = 1.65, outlier = 0.001)
}
\arguments{
\item{bmat}{sparse matrix, feature by cells, NOTE: in SnapATAC, bmat is cell by feature}

\item{nFragms}{data.frame, cols: barcode, nFragment}
}
\value{
list, bmat: updated bmat; rows: kept rows in original bmat
}
\description{
\itemize{
\item Filter features if the covariances are smaller or larger than thresholds.
\item Too high values will be set as zeros.
}
}
