% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{SnapATAC_runDiffusionMapsExtension}
\alias{SnapATAC_runDiffusionMapsExtension}
\title{Run DiffusionMaps for query cells baesd on the landmark results.}
\usage{
SnapATAC_runDiffusionMapsExtension(bmatLandmark, bmatQuery, mapLandmark)
}
\arguments{
\item{bmatLandmark}{matrix, cell by feature}

\item{bmatQuery}{matrix, cell by feature}

\item{mapLandmark}{results from difusionMaps for landmark cells.}
}
\value{
list of two elements
\itemize{
\item dmat matrix, nquery cell by feature
\item sdev numeric vector, standard deviation for the query cells.
}
}
\description{
Run DiffusionMaps for query cells baesd on the landmark results.
}
