% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityControl.R
\name{sumFragmentSingleThread}
\alias{sumFragmentSingleThread}
\title{get barcodes-related fragments stats.}
\usage{
sumFragmentSingleThread(
  tabixFile,
  chromSizes,
  outdir,
  sampleName = NULL,
  barcodes = NULL,
  nChunk = 3
)
}
\arguments{
\item{nChunk}{integer partition each chrom into nChunk pieces}
}
\description{
Ref: ArchR
}
